Rename the snap to minio-client (#2243)
In linux, mc is midnight commander. In the traditional
linux distro world, name conflicts are forbidden. Because
we are trying to make this decentralized and self-serving,
we need less strict rules. For this, snap has a concept
of aliases. Then the user can decide what application owns
the mc name.