fs: Simply PUT and fix issue when the source is not seekable
1. Functional change:

In any case, the downloaded data will be stored in [file].part.minio
first and then renamed to [file]

2. Bug fixing:

The code also does not behave correctly when the source reader is not
seekable. reader is wrapped by a reader hook and then tested if it is
seekable or not, though the reader hook is always seekable.