fix copy issue when the source is not seekable (#3216)
The downloaded data will be stored in [file].part.minio
first and then renamed to [file]

The code also does not behave correctly when the source 
reader is not seekable. the reader is wrapped by a reader 
hook and then tested if it is seekable or not, though the 
reader hook is always seekable.