Implement new format of the health report
To ensure compatibility with older versions of `mc`, `minio` server will continue
to return the health info in the current format. `mc` will internally transform/map
this data to the new data structure required to generate the output in the new
format. The new format is as follows:

 - hardware
   - servers
     - cpus
     - meminfo
     - perf
       - net
       - drives
 - software
  - minio
    - info
      - servers
      - drives
    - config
    - procinfos
  - osinfos

The new format also includes a header at the top of the report that provides
information about the version of the report schema, and looks like:

```json
{
    "subnet": {
      "health": {
        "version": "v1"
      }
    }
  }
```