Remove IsInstalled() call before auto-completion installation
At the moment posener/complete [1] doesn't allow specifying shell for
IsInstalled(), which might lead to confusing error messages, since it
just returns true if it's able to find completion definition in *any*
shell configurations even if it's not what the user currently uses.

Hence we should remove this check and call Install() directly, and user
can get more details in this case.

[1] https://github.com/posener/complete