cp: Copy recursive paths having the passed prefix
Some users used to copy a subset of files stored under one specific prefix
using `mc cp -r alias/bucket/[starting-characters] /dest`.

However this is not working anymore.

The reason is that before, let's say if we have a prefix containing
an object called '1.txt', client-s3.Stat() passed with a prefix '1'
will return the info of the object '1.txt'. However this behavior
was changed recently and client-s3.Stat() now only returns the
information of the exact object name '1'.

To solve this, we add firstURL2Stat() which returns the stat info of the
first object having the specified prefix.