Add option to upload health report to Subnet
Add an optional text flag `license` to the `mc admin subnet health`
command. If this flag is passed, upload the genearated health report to
Subnet by passing the license key as auth.

For use in development mode (where the report needs to be uploaded to
Subnet running on localhost), introduce another optional (and hidden)
boolean flag called `dev`. If it is passed, use `http://localhost:9000`
as the base url for Subnet instead of `https://subnet.min.io`

This feature can subsequently be used to automate the health report
upload by either using a cron job or introducing another `frequency`
flag to the command.