Add --name flag to subnet health command
If passed along with the --license flag, this will be passed as the name
of the cluster in the subnet upload request. Subnet uses this only for
the 1st upload of the given cluster.

If not passed, the alias of the cluster will be passed as default value
for the cluster name.