Fix incorrect error related to "name" flag
The command `mc admin subnet health alias` is throwing an error saying
"--name is applicable only when --license is also passe" even when the
flag `--name` is not passed.

This is happening because the name is default-populated with the alias
passed in the command. Fixed my moving the default-population after the
validation.