fix: health report incompatibility with old minio
When the subnet health command from current mc is used with the MinIO
versions before the latest health data structure changes (v2) were
introduced, the generated report doesn't contain the cluster information
(minio -> info).

This is because these old versions of minio don't return this
information in the output of the `healthinfo` api.

Fixed this by explicitly pulling the server info separately when we
detect that the health schema version is not v2.