Fix incorrect spinner for network perf test
In case of non-distributed setup, the server doesn't return network perf
data. Because of this, the network spinner doesn't finish, and is closed
explicitly outside the loop, resulting two lines (one incomplete, one
complete) for it.

Fixed by explicitly checking the number of servers when checking for
completion of the network spinner.