Use subnet api key instead of license
The earlier approach of using a license token for communicating with
SUBNET is being replaced with a simpler mechanism of api keys. Unlike
the license which is a JWT token, these api keys will be simple UUID
tokens, and don't have any embedded information in them. SUBNET would
generate the api key on cluster registration, and return in the api
response. mc then stores it in the `subnet api_key` config of minio,
to be used for subsequent communication with SUBNET.

In case the api_key is not available in config, but the cluster's
license is available, it will be used as auth in the subnet api call,
and api_key will be set after the call succeeds. This will help in
smooth migration from older versions of minio.