Get speedtest results as part of `support diag`
Also remove the `--full` flag and run the new perf tests (that run quite
fast) always (if the minio version supports it).

In case of older versions of minio that have the older versions of the
performance tests, mc will not request for those tests to be run, and
subsequently will not look for their results in the response.