alias: fetch peer certificate when adding an alias
This commit adds an extension to the `alias set`
command. When adding an alias fails due to a TLS
trust error - i.e. remote peer certificate verification
fails - `mc` fetches the remote peer certificate,
computes the public key fingerprint and asks the user
to confirm the fingerprint.

When the user confirms the fingerprint, the certificate
is added to the `CAs` directory.

This allows users to use a TOFU authentication with
self-signed certificates - similar to SSH.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>