Updating the language to be more clear
When I read this the first time, I thought it was saying it would remove all objects in the bucket and the bucket itself ("remove bucket"). I did NOT understand it to mean that it would remove ALL buckets in the account and so cheerfully added the force and dangerous flags!!

Fortunately, they were all test buckets, but can you imagine if I had run this on a production account? Yikes! So this is a simple patch to just change the language a little bit to make it more obvious. If a non-native English speaker can think of a simpler way to word this, even better.