New offline registration flow
Instead of asking the user to
- open the subnet register page
- paste the register token
- copy the generated license token
- paste it back on mc

Now we will simply ask the user to visit a subnet URL (which contains
the registration token as a query parameter). On visiting this page,
subnet will automatically register the cluster and generate a license
for it.

User can then optionally update the license on mc using the
`mc license update` command