Upload results of perf tests to SUBNET
Apart from displaying the summary of the perf results on the screen,
generate a zip file containing the cluster metadata file (cluster.info)
and a JSON file containing the perf test results, and upload it to
SUBNET. Create this file as a temp file so that it gets automatically
deleted.

When the `--airgap` flag is passed (or if SUBNET upload fails), move the
temp file to current directory so that it can be inspected manually.