allow terminal control back after bubbletea UI
terminal control was not relinquished since ui.Start()
must not be run in a go-routine as it is expected
to be a blocking call, and must be serialized with when
an application exits.

refactor the 'scanner info/status' to ensure we have
correct behavior.

This PR also renames 'info-> status', keeps the older
one as alias for now.