Improve format of perf tests output
The current JSON format of the perf tests output is same as what mc
receives in response from the minio api. This is tailored more for
effectively rendering the real-time output on screen.

In order to retain this existing behavior, and to ensure backward
compatibility with older versions of minio, we will transform the minio
response into a more readable format to be used for uploading to subnet
and printing when `--json` is passed.

Following are the changes made:

- Remove unnecessary fields:
  - final
  - type
  - version
- Remove `ttfb` from `PUTStats`
- Rename `throughputPerSec` to `throughput`
- Rename `disks` to `drives`
- Rename `err` to `error`