Mint Suite/MC: Enable SSE-S3 testing
NOTE: To enable SSE-S3 env var the change on mint fork is also necessary!

Since we using SSE-S3 for encryption at rest we have to test it. With this change we're able to set SSE-S3 in the test setup if the env var is set to 1. Then all mc related tests of the mint suite will run against encrypted buckets. Tested it successfully during QA for 1.12.0, but onlx locally. Jenkins run is still to do.