Improve message when profile upload fails
When profile upload to SUBNET fails because of any error, currently just
the error message is shown, creating an impression that the profile file
was not saved anywhere. (we delete the file only if upload goes through,
otherwise it is stored locally)

Improve the message(s) printed on terminal in case of such failure,
clearly indicating that the file has been stored locally.