ls: Fix slow listing when an object has astronomical number of versions
mc ls can look like hang when it tries to list a prefix or an object
which has too many number of versions. Though the server returns the
listing quickly, mc delays showing the output because it waits for all
versions of each object to show the ordinal number of a version.

This can cause out of memory issue as well for machines with low memory.

The only way to fix this is to revert the ordinal number of versions:
- latest version will be v1
- older versions will be vN

This is a breaking change.