stat: Fix printing bucket usage info
A recent regression was introduced after refactoring some code related
to mc stat command but the usage is not correctly showing anymore. The
reason is that the new code relies on bucket info key field to get the
bucket stats from the data usage. This commit fixes the behavior by
populating Key field in bucket info structure.