heal: Fix summary ETA with multiple disks healing in one set
The short healing report calculates the ETA of finished healing based on
the disk that has the most healed objects in it. However one user can
replace another disk in the same set after healing a bunch of objects
in the first disk, which can lead to wrong ETA calculation.

Fix it by calculating the furthest ETA in the set.