Added `--recursive` option for tag removal
Earlier there was an issue where with `--versions` option
where the tags were removed for the bucket level again and
again, no of subdirs times under the bucket. This PR fixes
the issue and also introduces `--recursive` option for
removing the tags for objects/versions of subdirs recursively
(if need be).

Signed-off-by: Shubhendu Ram Tripathi <shubhendu@minio.io>