Show groups with their policy details while listing users
Currently while listing users there is no way to find out what policies
are being inherited due to group membership of the user. We need to get
details of group to map the details.
This change enables listing of user with their group having policies
list as well in `--json` output.

Signed-off-by: Shubhendu Ram Tripathi <shubhendu@minio.io>