revamp spinner in `mc support diag`
The current spinner mechanism looks for the data categories (CPU,
drives, memory, etc) sequentially. If any of the category's data is not
received, it gets stuck there.

If a new data category is introduced without changing the version of the
HealthInfo struct, then `mc support diag` can get stuck for the new
category when talking to an older version of minio.

So instead of looking for the categories one by one, start a spinner for
each category at the beginning, and mark it as done as soon as the
corresponding data is received. When the last category (admin info) is
received, mark any pending category as done so that the command
continues with rest of the processing.