Support metrics-v3 api in `admin prometheus metrics`
Add a flag --api-version with possible values of v2 and v3, with default
being v2. So when this flag is not passed, the command will work exactly
as it did before.

When using v3, there is a different set of metric types supported. It
also supports two additional flags:

bucket: the bucket for which the metrics are to be fetched. it is
applicable only for the metric types that are available at bucket level,
which currently are 'api' and 'replication'

list: when passed, this will only print the list of all metrics
available for the given metrics type (without fetching their values)