scanner: Add support of bucket per ES scan report
mc admin scanner status --bucket <bucket-name> ALIAS

will print detailed report of a particular bucket scan.

Like, when the last scan progress was saved for that bucket for all
erasure sets. Also it will print when a full scan of a bucket finished.

(A full bucket scan means the last time frame that was needed for each
of erasure sets to do sixteen cycles)