Do not fail mc-admin-policy-attach if policy already attached/detached
Currently, attempts to attach a policy to a user who already has
the policy attached to them results in a 400.

This change is to handle cases where policy attach/detach operations
are automated using scripts/jobs. A re-run of the attach/detach
step should not result in the failure of the entire job.