heal: Simplify bucket healing status
Avoid using drive count concept to pick a color for a bucket healing
result. This was straighforward when we were healing a bucket per erasure set
but now it becomes more complex to show a result.

Pick green if everything is okay, yellow if a bucket is missing in a
drive, red for any other drive state (unformatted, offline, ..) and
grey for any weird situation.